refactor: move guest auth to data access layer with server-only guard
